# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API to interface with Sidero Omni.",
        "title": "Talos Omni Control API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/infra-machine-configs": {
            "get": {
                "description": "Get a list of all infrastructure machine configs in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inframachineconfigs"
                ],
                "summary": "List all infrastructure machine configs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by machine ID",
                        "name": "machine",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.InfraMachineConfigResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/infra-machine-configs/{id}": {
            "get": {
                "description": "Get detailed information about a specific infrastructure machine config",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inframachineconfigs"
                ],
                "summary": "Get a single infrastructure machine config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Infrastructure Machine Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.InfraMachineConfigResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/installation-medias": {
            "get": {
                "description": "Get a list of all installation medias in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "installationmedias"
                ],
                "summary": "List all installation medias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.InstallationMediaResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/installation-medias/{id}": {
            "get": {
                "description": "Get detailed information about a specific installation media",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "installationmedias"
                ],
                "summary": "Get a single installation media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Installation Media ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.InstallationMediaResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/kernel-args": {
            "get": {
                "description": "Get a list of all kernel args configurations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "List kernel args",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.KernelArgsResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/kernel-args/{id}": {
            "get": {
                "description": "Get detailed information about a specific kernel args configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Get kernel args",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kernel Args ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.KernelArgsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/machines": {
            "get": {
                "description": "Get a list of all machines in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "List all machines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.MachineResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.InfraMachineConfigResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "acceptance_status": {
                    "type": "string"
                },
                "cordoned": {
                    "type": "boolean"
                },
                "extra_kernel_args": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "power_state": {
                    "type": "string"
                },
                "requested_reboot_id": {
                    "type": "string"
                }
            }
        },
        "handlers.InstallationMediaResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "architecture": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "dest_file_prefix": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "min_talos_version": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "no_secure_boot": {
                    "type": "boolean"
                },
                "overlay": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "src_file_prefix": {
                    "type": "string"
                }
            }
        },
        "handlers.KernelArgsResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
        "handlers.MachineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "arch": {
                    "type": "string"
                },
                "connected": {
                    "type": "boolean"
                },
                "hostname": {
                    "description": "Status fields (from MachineStatus resource)",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "last_error": {
                    "type": "string"
                },
                "maintenance": {
                    "type": "boolean"
                },
                "management_address": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "talos_version": {
                    "type": "string"
                },
                "use_grpc_tunnel": {
                    "type": "boolean"
                }
            }
        },
...
    }
}